requirements.txt
requests==2.31.0
redis==5.0.1
mysql-connector-python==8.3.0
beautifulsoup4==4.12.3

